refactor: 헤더 구조 개선 및 네이밍 변경 (#203)
- 메뉴 리스트 NavigationDrawer로 분리
- 네이밍 변경: HeaderMenu -> AppHeader
