+++ src/components/Common/index.ts
@@ -1,4 +1,4 @@
-export * from './HeaderMenu';
+export * from './Header/AppHeader';
 export * from './RootLayout';
 export * from './TopNavigator';
 export * from './Button';
